Explicitely convert an arg to the vector<unsigned int> constructor to avoid that it tries to take the constructor that uses two input iterators.
git-svn-id: https://svn.dealii.org/trunk@4793 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/fe/fe_q.cc
@@ -713,7 +713,7 @@ template <int dim>
 std::vector<unsigned int>
 FE_Q<dim>::get_dpo_vector(const unsigned int deg)
 {
-  std::vector<unsigned int> dpo(dim+1, 1);
+  std::vector<unsigned int> dpo(dim+1, static_cast<unsigned int>(1));
   for (unsigned int i=1; i<dpo.size(); ++i)
     dpo[i]=dpo[i-1]*(deg-1);
   return dpo;
